DipSnipe/requirements.txt
flask
flask-cors
psycopg2-binary
requests
pandas
numpy
yfinance
pyyaml